+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Adding support for `latitude` and `longitude`.
+
 ## [0.2.0] - 2021-05-12
 
 ### Changed